MLmonitoringapp/requirements.txt
azure-functions
psycopg2-binary>=2.9.9
mlflow>=2.10.0
transformers>=4.37.0
torch>=2.2.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
python-dotenv>=1.0.0
azure-storage-blob>=12.14.0
azure-storage-queue>=12.6.0
azure-identity>=1.12.0
python-dateutil>=2.8.2
pandas>=1.5.0
scikit-learn>=1.0.2
matplotlib>=3.5.0
numpy>=1.22.0
mlflow>=2.0.0
azureml-core>=1.48.0